requirements.txt
discord.py>=2.0,<3
python-dotenv>=1.0,<2
psycopg2-binary>=2.9,<3
aiohttp>=3.9,<4
